13-section-integration-testing/src/app/routing/routing.component.ts
import { Component, OnInit } from '@angular/core';
import { ActivatedRoute, Params, Router } from '@angular/router';

@Component({
  selector: 'app-routing',
  templateUrl: './routing.component.html',
  styleUrl: './routing.component.scss'
})
export class RoutingComponent implements OnInit {
  constructor(
    private router: Router,
    private route: ActivatedRoute
  ) {}

  ngOnInit(): void {
    this.route.params.subscribe((params: Params): void => {
      if (params['id'] === '0') {
        this.router.navigate(['/404']);
      }
    });
  }

  goBack(): void {
    this.router.navigate(['/posts']);
  }
}
